refactor: Extract getBestThumbUrl into helper
